Add an account check to the CI workflow
gh-13

I have this issue of keeping track of which account and region my
various applications are deployed to. I'd like to converge on keeping
the AWS account ID and region saved in GitHub Actions Variables, and
asserting in Actions that the account we wind up authenticated to
matches what's recorded in Variables.